refactor: renames the API modules folder from 'route' to 'api'
