+++ src/server.ts
@@ -3,13 +3,13 @@ import express, { Express } from 'express';
 import helmet from 'helmet';
 import { pino } from 'pino';
 
+import { healthCheckRouter } from '@/api/healthCheck/healthCheckRouter';
+import { userRouter } from '@/api/user/userRouter';
 import { openAPIRouter } from '@/api-docs/openAPIRouter';
 import errorHandler from '@/common/middleware/errorHandler';
 import rateLimiter from '@/common/middleware/rateLimiter';
 import requestLogger from '@/common/middleware/requestLogger';
 import { env } from '@/common/utils/envConfig';
-import { healthCheckRouter } from '@/routes/healthCheck/healthCheckRouter';
-import { userRouter } from '@/routes/user/userRouter';
 
 const logger = pino({ name: 'server start' });
 const app: Express = express();
